+++ src-docs/src/views/form/form_popover.js
@@ -12,9 +12,7 @@ import {
   EuiFieldText,
 } from '../../../../src/components';
 
-function makeId() {
-  return Math.random().toString(36).substr(2, 5);
-}
+import makeId from '../../../../src/components/form/form_row/make_id';
 
 export default class extends Component {
   constructor(props) {
@@ -69,14 +67,12 @@ export default class extends Component {
         </EuiFormRow>
 
         <EuiFormRow
-          id={makeId()}
           label="A text field"
         >
           <EuiFieldText name="popfirst" />
         </EuiFormRow>
 
         <EuiFormRow
-          id={makeId()}
           label="Range"
           helpText="Some help text for the range"
         >