+++ src-docs/src/views/form/inline_form_sizing.js
@@ -9,21 +9,15 @@ import {
   EuiFieldNumber,
 } from '../../../../src/components/';
 
-function makeId() {
-  return Math.random().toString(36).substr(2, 5);
-}
-
-const idPrefix = makeId();
-
 export default () => (
   <EuiFlexGroup style={{ maxWidth: 600 }}>
     <EuiFlexItem grow={false} style={{ width: 100 }}>
-      <EuiFormRow label="Age"  id={idPrefix}>
+      <EuiFormRow label="Age">
         <EuiFieldNumber max={10} placeholder={42} />
       </EuiFormRow>
     </EuiFlexItem>
     <EuiFlexItem>
-      <EuiFormRow label="Full name" id={idPrefix}>
+      <EuiFormRow label="Full name">
         <EuiFieldText icon="user" placeholder="John Doe" />
       </EuiFormRow>
     </EuiFlexItem>